+++ apps/condo/pages/auth/input.tsx
@@ -1,0 +1,9 @@
+import { InputPhoneForm } from '@condo/domains/user/components/auth/InputPhoneForm'
+import { RegisterLayout } from '@condo/domains/user/components/auth/RegisterLayout'
+
+export default function InputPage () {
+    return (
+        <InputPhoneForm />
+    )
+}
+InputPage.container = RegisterLayout
