+++ apps/condo/pages/auth/register.tsx
@@ -1,39 +1,25 @@
-import { Button } from '@condo/domains/common/components/Button'
-import { BasicEmptyListView } from '@condo/domains/common/components/EmptyListView'
-import { InputPhoneForm } from '@condo/domains/user/components/auth/InputPhoneForm'
-import { RegisterContext, RegisterContextProvider } from '@condo/domains/user/components/auth/RegisterContextProvider'
+import { getClientSideSenderInfo } from '@condo/domains/common/utils/userid.utils'
 import { RegisterForm } from '@condo/domains/user/components/auth/RegisterForm'
-import { ValidatePhoneForm } from '@condo/domains/user/components/auth/ValidatePhoneForm'
-import AuthLayout, { AuthPage } from '@condo/domains/user/components/containers/AuthLayout'
-
-import { useIntl } from '@core/next/intl'
-import { Col, Row, Typography } from 'antd'
-import Router from 'next/router'
-import React, { useContext, useEffect, useState } from 'react'
-import { ButtonHeaderAction } from '@condo/domains/common/components/HeaderActions'
+import { useRouter } from 'next/router'
+import { useCallback, useEffect } from 'react'
 import { useMutation } from '@core/next/apollo'
 import { runMutation } from '@condo/domains/common/utils/mutations.utils'
-import { getClientSideSenderInfo } from '@condo/domains/common/utils/userid.utils'
+import { useIntl } from '@core/next/intl'
 import { CREATE_ONBOARDING_MUTATION } from '@condo/domains/onboarding/gql'
-import { fontSizes } from '@condo/domains/common/constants/style'
+import { RegisterLayout } from '@condo/domains/user/components/auth/RegisterLayout'
+import { useRegisterContext } from '@condo/domains/user/components/auth/RegisterContextProvider'
 
-const RegisterPage: AuthPage = () => {
+export default function RegisterPage () {
     const intl = useIntl()
-    const RegistrationTitleMsg = intl.formatMessage({ id: 'pages.auth.RegistrationTitle' })
-    const PhoneConfirmTokenErrorLabel = intl.formatMessage({ id: 'pages.auth.register.PhoneConfirmTokenErrorLabel' })
-    const PhoneConfirmTokenErrorMessage = intl.formatMessage({ id: 'pages.auth.register.PhoneConfirmTokenErrorMessage' })
-    const RestartPhoneConfirmLabel = intl.formatMessage({ id: 'pages.auth.register.RestartPhoneConfirmLabel' })
-
-    const { token, isConfirmed, tokenError, setToken, setTokenError } = useContext(RegisterContext)
-    const [state, setState] = useState('inputPhone')
-
+    const { token, tokenLoading, isConfirmed, setToken } = useRegisterContext()
+    const router = useRouter()
     const [createOnBoarding] = useMutation(CREATE_ONBOARDING_MUTATION, {
         onCompleted: () => {
-            Router.push('/onboarding')
+            router.push('/onboarding')
         },
     })
 
-    const initOnBoarding = (userId: string) => {
+    const initOnBoarding = useCallback((userId: string) => {
         const onBoardingExtraData = {
             dv: 1,
             sender: getClientSideSenderInfo(),
@@ -41,80 +27,26 @@ const RegisterPage: AuthPage = () => {
 
         const data = { ...onBoardingExtraData, type: 'ADMINISTRATOR', userId }
 
-        return runMutation({
+        runMutation({
             mutation: createOnBoarding,
             variables: { data },
             intl,
         })
-    }
+    }, [createOnBoarding, intl])
+
+    const onRegisterFinish = useCallback((userId: string) => {
+        setToken(null)
+        initOnBoarding(userId)
+    }, [initOnBoarding, setToken])
 
     useEffect(() => {
-        if (token && isConfirmed) {
-            setState('register')
-        } else if (token) {
-            setState('validatePhone')
-        } else {
-            setState('inputPhone')
+        if (!tokenLoading && (!token || !isConfirmed)) {
+            router.push('/auth/input')
         }
-    }, [token, isConfirmed])
-
-    if (tokenError && token) {
-        return (
-            <BasicEmptyListView>
-                <Typography.Title level={3}>
-                    {PhoneConfirmTokenErrorLabel}
-                </Typography.Title>
-                <Typography.Text style={{ fontSize: fontSizes.content }}>
-                    {PhoneConfirmTokenErrorMessage}
-                </Typography.Text>
-                <Button
-                    type='sberPrimary'
-                    style={{ marginTop: '16px' }}
-                    onClick={() => {
-                        setToken(null)
-                        setTokenError(null)
-                        setState('inputPhone')
-                        Router.push('/auth/register')
-                    }}
-                >
-                    {RestartPhoneConfirmLabel}
-                </Button>
-            </BasicEmptyListView>
-        )
-    }
-
-    const steps = {
-        inputPhone: <InputPhoneForm onFinish={() => setState('validatePhone')} />,
-        validatePhone: <ValidatePhoneForm
-            onFinish={() => setState('register')}
-            onReset={() => {
-                setState('inputPhone')
-                Router.push('/auth/register')
-            }}
-        />,
-        register: <RegisterForm onFinish={initOnBoarding} />,
-    }
+    }, [tokenLoading, token, isConfirmed, router])
 
     return (
-        <RegisterContextProvider>
-            <Row gutter={[0, 24]}>
-                <Col span={24}>
-                    <Typography.Title>{RegistrationTitleMsg}</Typography.Title>
-                </Col>
-                <Col span={24}>
-                    { steps[state] }
-                </Col>
-            </Row>
-        </RegisterContextProvider>
+        <RegisterForm onFinish={onRegisterFinish} />
     )
 }
-
-RegisterPage.headerAction = (
-    <ButtonHeaderAction
-        descriptor={{ id: 'pages.auth.AlreadyRegistered' }}
-        path={'/auth/signin'}
-    />
-)
-RegisterPage.container = AuthLayout
-
-export default RegisterPage
+RegisterPage.container = RegisterLayout
