+++ apps/condo/pages/auth/signin.tsx
@@ -27,7 +27,7 @@ const SignInPage: AuthPage = () => {
     )
 }
 
-SignInPage.headerAction = <ButtonHeaderAction descriptor={{ id: 'pages.auth.Register' }} path={'/auth/register'}/>
+SignInPage.headerAction = <ButtonHeaderAction descriptor={{ id: 'pages.auth.Register' }} path={'/auth/input'}/>
 
 SignInPage.container = AuthLayout
